Completed Registration/Authefication/error404/error403

+++ mvc/application/controllers/PostController.php
@@ -7,21 +7,30 @@ class PostController extends Controller
 
     public function create()
     {
-        $this->view->render('post/create');
+        if (!empty($_SESSION['email'])) {
+            $this->view->render('post/create');
+        } else {
+            $this->view->render('auth/403');
+        }
     }
     public function save()
     {
-        $data = $_POST;
-        $data['created_at'] = date('Y-m-d H:m:s');
-        if (!empty($data)) {
-            $this->getModel('post');
-            $this->model->create($data);
+        if (!empty($_SESSION['email'])) {
+            $data = $_POST;
+            $data['created_at'] = date('Y-m-d H:m:s');
+            if (!empty($data)) {
+                $this->getModel('post');
+                $this->model->create($data);
 
-            if(isset($_SERVER['HTTP_REFERER'])) {
-                $main = $_SERVER['HTTP_ORIGIN'];
-                header('Location:' . $main);
+                if (isset($_SERVER['HTTP_REFERER'])) {
+                    $main = $_SERVER['HTTP_ORIGIN'];
+                    header('Location:' . $main);
+                }
             }
+        } else {
+            $this->view->render('auth/403');
         }
+
     }
     public function show($id)
     {
@@ -32,27 +41,39 @@ class PostController extends Controller
 
     public function update($id)
     {
-        $this->getModel('post');
-        $post = $this->model->findById($id);
-        $this->view->render('post/update',$post);
+        if (!empty($_SESSION['email'])) {
+            $this->getModel('post');
+            $post = $this->model->findById($id);
+            $this->view->render('post/update',$post);
+        } else {
+            $this->view->render('auth/403');
+        }
     }
 
     public function edit($id)
     {
-        $data = $_POST;
-        if (!empty($data)) {
-            $this->getModel('post');
-            $this->model->update($data, $id);
-            $main = $_SERVER['HTTP_ORIGIN'];
-            header('Location:' . $main);
+        if (!empty($_SESSION['email'])) {
+            $data = $_POST;
+            if (!empty($data)) {
+                $this->getModel('post');
+                $this->model->update($data, $id);
+                $main = $_SERVER['HTTP_ORIGIN'];
+                header('Location:' . $main);
+            }
+        } else {
+            $this->view->render('auth/403');
         }
     }
 
     public function delete($id)
     {
-        $this->getModel('post');
-        $this->model->delete($id);
-        header('Location:/');
+        if (!empty($_SESSION['email'])) {
+            $this->getModel('post');
+            $this->model->delete($id);
+            header('Location:/');
+        } else {
+            $this->view->render('auth/403');
+        }
     }
 }
